fix: dashboard filter > circular type dependency in core
