+++ packages/dashboard/src/components/DashboardFilters/DashboardFiltersWrapper.tsx
@@ -5,7 +5,7 @@ import { changeFilterActive } from '../../helpers/changeFilterActive'
 import _ from 'lodash'
 import { FilterBehavior } from '../../helpers/FilterBehavior'
 import { getFilteredData } from '../../helpers/getFilteredData'
-import { DashboardFilters } from '../../types/DashboardFilters'
+import { type DashboardFilters } from '@cdc/core/types/DashboardFilters'
 import { getQueryParams, updateQueryString } from '@cdc/core/helpers/queryStringUtils'
 import { VisualizationWrapper, Sidebar, Responsive } from '@cdc/core/components/Layout'
 import DashboardFiltersEditor from './DashboardFiltersEditor'
